backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv
requests
psycopg2-binary
bcrypt
scikit-learn>=1.2.0
joblib==1.4.2
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
python-jose[cryptography]
